[202412] Add pipeline check for missing __init__.py in sonic-xcvrd whl package (#593)
Description
Ensure that the pipeline fail to build if sonic-xcvrd whl package is missing init.py file.

Motivation and Context
The __init__.py file is necessary in each directory to indicate that the directory should be treated as a package.
Tools like setuptools and find_packages() rely on the presence of __init__.py files to identify packages and subpackages.
Without __init__.py files, the directories will not be included in the package distribution, leading to incomplete or broken packages.
By ensuring that each directory contains an __init__.py file, we can create a complete and functional Python wheel package.